+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :rocket: Features
 
+* feat(metrics): add experimental advisory attributes support
 ### :bug: Bug Fixes
 
 ### :books: Documentation